\name{elemental.formula}
\alias{elemental.formula}

\title{
List from elemental formula
}
\description{
Obtain a list from elemental formula given in the database. Use internally by the Isotopic.R function. Adaptation from R-package OrgMassSpecR.
}
\usage{elemental.formula(elemental.formula)
}

\arguments{
  \item{elemental.formula}{
Elemental formula}
  }

\author{Manuel D Peris Diaz}
\references{
1. R-MetaboList: a flexible tool for metabolite extraction from high-resolution data-independent acquisition mass spectrometry analysis. Metabolites. Soon

2. A Survey of Orbitrap All Ion Fragmentation Analysis Assessed by an R MetaboList Package to Study Small-Molecule Metabolites. Chromatographia. 2018, 81, 981-994.

}
\examples{
#library(MetaboList)
#Reading the library and indicating type of adduct
#Isotopic(library,name=c("M+H"))


}
